// web/lib/api/generated/model/movementTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `INWARD` - Goods received
 * * `SALE` - Dispatched
 * * `RETURN` - Returned
 * * `ADJUSTMENT_IN` - Adjustment in
 * * `ADJUSTMENT_OUT` - Adjustment out
 * * `DAMAGE` - Damaged
 * * `TRANSFER_IN` - Transfer in
 * * `TRANSFER_OUT` - Transfer out
 * * `RESERVE` - Reserved for an order
 * * `RELEASE` - Reservation released
 */
export type MovementTypeEnum = typeof MovementTypeEnum[keyof typeof MovementTypeEnum];


export const MovementTypeEnum = {
  INWARD: 'INWARD',
  SALE: 'SALE',
  RETURN: 'RETURN',
  ADJUSTMENT_IN: 'ADJUSTMENT_IN',
  ADJUSTMENT_OUT: 'ADJUSTMENT_OUT',
  DAMAGE: 'DAMAGE',
  TRANSFER_IN: 'TRANSFER_IN',
  TRANSFER_OUT: 'TRANSFER_OUT',
  RESERVE: 'RESERVE',
  RELEASE: 'RELEASE',
} as const;